moving endpoints to separate files and updating how we pass unread_message_total
